Add files via upload
## Music Genre Classification using Audio Analysis

This project implements a machine learning model for classifying music tracks into different genres based on audio analysis. It utilizes the Librosa library for extracting audio features and the scikit-learn and TensorFlow frameworks for building and training the classification model.

The code scrapes the Last.fm website to obtain a dataset of music tracks and their corresponding genres. It then downloads the audio files and extracts features such as chroma_stft, spectral_centroid, and mfcc. These features are used to train a deep learning model, which is capable of predicting the genre of a given music track.

The project showcases the application of machine learning techniques in the field of music analysis and genre classification. By leveraging audio processing and machine learning algorithms, it provides insights into the potential of using audio features for automated genre classification tasks.

Feel free to explore the code and experiment with different datasets to further improve the model's performance. Any feedback and contributions are welcome!